Moved the cache into its own package
