+++ pkg/credentials/credentials_suite_test.go
@@ -1,0 +1,11 @@
+package credentials_test
+
+import (
+	"testing"
+
+	"github.com/weaveworks/eksctl/pkg/testutils"
+)
+
+func TestCredentials(t *testing.T) {
+	testutils.RegisterAndRun(t)
+}
