+++ pkg/eks/api.go
@@ -43,6 +43,7 @@ import (
 	api "github.com/weaveworks/eksctl/pkg/apis/eksctl.io/v1alpha5"
 	"github.com/weaveworks/eksctl/pkg/az"
 	"github.com/weaveworks/eksctl/pkg/cfn/manager"
+	ekscreds "github.com/weaveworks/eksctl/pkg/credentials"
 	"github.com/weaveworks/eksctl/pkg/kubernetes"
 	kubewrapper "github.com/weaveworks/eksctl/pkg/kubernetes"
 	"github.com/weaveworks/eksctl/pkg/utils"
@@ -161,9 +162,9 @@ func New(spec *api.ProviderConfig, clusterSpec *api.ClusterConfig) (*ClusterProv
 	// later re-use if overriding sessions due to custom URL
 	s := c.newSession(spec)
 
-	cache := os.Getenv(EksctlGlobalEnableCachingEnvName)
+	cache := os.Getenv(ekscreds.EksctlGlobalEnableCachingEnvName)
 	if s.Config != nil && cache != "" {
-		if cachedProvider, err := NewFileCacheProvider(spec.Profile, s.Config.Credentials, &RealClock{}); err == nil {
+		if cachedProvider, err := ekscreds.NewFileCacheProvider(spec.Profile, s.Config.Credentials, &ekscreds.RealClock{}); err == nil {
 			s.Config.Credentials = credentials.NewCredentials(&cachedProvider)
 		} else {
 			logger.Warning("Failed to use cached provider: ", err)
